json_unwrap: Fix undefined behavior
We need to ensure that the reference _json stays valid during usage.
When wrapping arguments to a function call, the lifetimes of temporaries
are extended. Lifetime extension would work for stack-based references.
This does *not* work for member variables that are references, though
(see https://herbsutter.com/2008/01/01/gotw-88-a-candidate-for-the-most-important-const/).
Hence, instead of binding the ref to a temporary Json() in
QObject::property(), define a static "null" JSON object to reference.