QObject::addSignal(): Don't replace notify signals with pure signals
When signals are overloaded, addSignal() might replace an existing signal.
If that existing signal is a property notify signal and the new one is not,
it will effectively eliminate the possibility of reacting to property
updates, since they come batched and delayed.
This patch prevents replacement of existing notify signals with pure
signals. Signals of the same kind can still be replaced to ensure that
actual property overrides continue to work.

Connecting to the overloaded pure signal is still possible by using the full
signal signature.